XML mappings - Refactor to move XML writing to a separate iteration
We have the potential for more parallel processing, and more importantly entity meta lookups, if we are aware of all entities at the time we begin XML generation. This allows associations on entity A to reference a property on entity B, and we can easily look up the entity name based on the class name.