Transactions - Implement basic transaction commit/rollback behavior
On calling transactionCommit() or transactionRollback(), commit or rollback the transaction (respectively) on all request datasources, then begin a new transaction.

Note we don't yet support transaction savepoints OR automatic transactions. (i.e. hibernate code isn't in a transaction by default, which is a major incompat we will need to address.)